add vector data visualization
