+++ .gitignore
@@ -26,6 +26,7 @@ wheels/
 *.egg-info/
 .installed.cfg
 *.egg
+docs/examples/data/
 
 # PyInstaller
 #  Usually these files are written by a python script from a template
