+++ examples/data/us_states.cpg
@@ -1,0 +1,1 @@
+UTF-8
